Finalized average visualization modules
Added comments and additional visualization code to finish off the average visualization code that we will implement after the user inputs their preferred genre and variables.  Added dynamic y scale to change with the variable and create a more accurate visualization as well as a horizontal line at the value of the chosen genre mean.